Rename component files to use their actual name
